requirements.txt
pyqt5
python-docx
faiss-cpu
numpy
pywin32